chore: move shadcn components to location more obvious to new users
